CogVm source as per VMMaker.oscog-eem.2032
Co/StackInterpreter:
Fix slip in makeBaseFrame:; the code to skip the primitive when activating a
method with a primitive used the wrong way to determine the initial PC of a
method.  Use the new startPCOfMethodHeader: (consistent with startPCOfMethod:).
Rename initialPCForHeader:method: to initialIPForHeader:method: to avoid
confusion; IP is used for instructionPointers, i.e. addresses of bytecodes,
not indexes of bytecodes.