Correct generateCogMethod/FullBlock which might write one byte too far.
Indeed, (padIfPossibleWithStopsFrom: startAddress to: endAddress) is inclusive (it will write a byte into endAddress).
So it must be called with (self padIfPossibleWithStopsFrom: startAddress to: startAddress + size - 1).

Maybe we were saved by roundUpLength:, or maybe the methodZone further allocated size was not yet used, but we don't have to live that dangerously.