VMMaker.oscog-eem.2057.mcz SmartSyntaxPluginSimulator: Fix several bugs with smart syntax operations (cPtrAsOop asSmallIntegerObj et al) that weren't being simulated.
Delete Object>>primitive: and implement it in SmartSyntaxInterpreterPlugin & SmartSyntaxPluginSimulator to be able to collect the signatures of zero-arg smart syntax prims.

Extend SmartSyntaxPluginSimulator instantiation to include a simulator subclass if one exists (the simulatorClass has to remain SmartSyntaxPluginSimulator to get the wrapping and marshalling on forwarding to work).

Fix returning self in SmartSyntaxPluginSimulator>>doesNotUnderstand:

Implement the ccgLoad:*andThen: methods.

FIx coercion of the byteArrayArg in the FilePluginSimulator read and write methods.