VMMaker.oscog-eem.2061.mcz Simulator: Use Levente's lock-free growth of the externalSemaphoreSignalRequests/Responses code.
Turn off logging in the SocketPluginSimulator now that it can do an update.

SpurPlanningCompactor:
Add an assert to check that all marked and unpinned objects in the mobile range are actually mobile, and hence uncover the bug in planCompactSavingForwarders.

Fix the bug in planCompactSavingForwarders (the rest to follow).

Add an interestingObject and halt the enumerations at it for debugging (simulation only).