SpurPlanningCompactor:
Add tests for pinned objects over which, and/or within which live objects are moved.

Move the emptyVM into SpurPlanningCompactorTestsImageResource as SUnit intends.

Fix the compaction enumeration.  The previousPin must not be reset until the enumeration is unable to find room for the current mobile object below previousPin.  So previousPin is only updated when a mobile object that won't fit beneath it is in hand.  Since this means that the size of the object in hand will potentially be tested multiple times, pull it out of the loop over pinned object from previousPin up, and hence refactor copyAndUnmarkObject:to:firstField: to copyAndUnmarkObject:to:bytes:firstField:, and have all the enumerations update toFinger explicitly.

Makew sure that pinned objects above the lastMobileObject are unmarked.  Pull this code into its own method.