SistaCogit: Fix conditional branches following inline primitive comparisons given the branch following logic.  The unconditional branch following can only be elided if jumping to the bytecode following the conditional branch, which mnay no longer be the case with branch following.  To this end refactor nextDescriptorAndExtensionsInto: into nextDescriptorExtensionsAndNextPCInto:.
SistaCogit Simulation:
Fix stack depth calculations for trap bytecodes and non-local returns in full blocks.

Fix frame pointer checking in handleCallOrJumpSimulationTrap: for primitiveSlotAtPut sent to a Context.

Neaten [detailed]symbolicMethod: eliding the VMProxy noise.

Fix Character printing in shortPrint: to add the $ in front of the character, not just the code.

Make debugBytecodePCs and debugOpcodeIndices specific to breakMethod if it is set.

Send Label to the right dude in noneImmediateBranchIf:notInstanceOfBehaviors:target:.

Can now simulate Clément's Scorch image enough to see the UI (FFI calls are not simulated so no getenv: hence no sources.  And a createDirectory: causes a primitive failed which shows up in the debugger in the simulation.)