CogVM source as per VMMaker.oscog-eem.2134
Move the SpurPlanningCompactor to tip.  It is now ready for prime time.
SpurPlanningCompactor is a three-finger algorithm that slides objects down to
the first finger, which points at the first free object, from the first
unpinned, marked object, while stepping around intervening pinned objects
pointed at by the third finger.

StackInterpreter:
Fix serious and ancient bug with stack checking and event checking.  The event
check piggy backs off the stack check, but the old code would not handle a stack
overflow if there was an overflow and an event check at one and the same time,
allowing frames to grow down the stack page in cases where there are lots of
events (such as profiling).

Fix assert fails in Spur due to failing perform primitives not restoring the
function pointer and argumentCount correctly.

Cogit:
Fix checking the lists of cog methods (openPICList and unpairedMethodList) when
compacting the method zone.
Fix a bug with freeing a method that might have either a newxtMethod or an IRC.
Must not free CogMethods.
Fix a couple of C compilation warnings.Have printCogMethods print totals of the
number of methods & PICs in the code zone.
Fix a slip in printCogMethodsWithPrimitive:.

Spur:
provide isUnmarked: for debugging with printOopsSuchThat:.
remember the first unmarked object found by allObjectsUnmarked in the bogon var,
and keep hold on to it via printBogons so it is not deleted by the C optimizer.