VMMaker.oscog-eem.2138 StackToRegisterMappingCogit: Reimplement mclassIsSmallInteger in terms of receiverTags, providing more generality at cheaper cost (receiverTags is computed during set-up for compiling a method).
Improve special-selector comparison and arithmetic based on this if the receiver or argument is self and known to be a SmallInteger.  Also improve the tag checking to avoid the spurious copy to TempReg unless both receiver and argument need to be tested.  These improve code quality in methods such as SmallInteger>>digitLength:.

Use anyMask: in ssAllocateRequiredRegMask:upThrough:upThroughNative: instead of bitAnd:...~= 0.

RegisterAllocatingCogit:
Override ssAllocateRequiredRegMask:upThrough:upThroughNative: to void optStatus and simStack entries that refer to the allocated register.

Simulator:
Aim stdio at the coInterpreter's transcript rather than Transcript.

Fix some tag-related bugs in the CurrentImageCoInterpreterFacade hierarchy.