CogVM source as per VMMaker.oscog-eem.2138
StackToRegisterMappingCogit:
Reimplement mclassIsSmallInteger in terms of receiverTags, providing more
generality at cheaper cost (receiverTags is computed during set-up for
compiling a method).

Improve special-selector comparison and arithmetic based on this if the receiver
or argument is self and known to be a SmallInteger.  Also improve the tag
checking to avoid the spurious copy to TempReg unless both receiver and
argument need to be tested.  These improve code quality in methods such as
SmallInteger>>digitLength:.

CoInterpreter:
Earlier assert checking for invalid class on send.

Address fortify issue in printf
printf(variable) is dangerous as variable can contain escape sequences. Change
the code to define print to printf("%s", s) as seen in other >>#print:
implementations.  This is necesary to allow packaging of VM in debian (they will
reject the source if it is not commited with this protection).
[eem: I'm committing this under protest.  The existing code is not unsafe and is
 simpler.  This is tyrrany by spurious warning.]

Adding a missing method that causes an undefined reference when building a 64
bits Lowcode VM.