unix: Simply check for the presence of mmap
There is little point to check if mmap is working on the system
that is building the software. Simply check if mmap is present
or not. This avoids running into issues around mmap_min_addr.

This is a backport of https://github.com/pharo-project/pharo-vm/pull/117/commits/43461db5c154d8c47f3c504a67e55b9890839c74
It is required on Ubuntu 14.04 (see step 2 in previous commit)
Other reference by the original author:

"I have not fully read the macro but I think the AC_FUNC_MMAP runs into the mmap_min_addr restriction. But as most people use a binary version of the VM there is little point to check that the system that built the software had a working version (and since the late 90s mmap generally works).

https://wiki.debian.org/mmap_min_addr"