VMMaker.oscog-eem.2144.mcz 64-bit Sista: Spur64BitMemoryManager must export fetchClassTagOf:. Spur: Check 64-bit translation for signed comparisons in SpurPlanningCompactor as well as 32-bit.
Cogit:
Rename blockPass, saveForBlockCompile et al to compilationPass, saveForRecompile etc.  This so that RegisterAllocatingCogit can recompile when discovering a loop that needs a merge so that the register set computed at the end of a loop can be installed at the head.
Refactor initializeFixupAt: into it and initializeFixup: to support RAC's recompilation.
Add some inlines to eliminate some thin intermediate functions.
Add support for adding #ifdef's to struct defs and use it to avoid CogSSBytecodeFixup's simNativeStackPtr and simNativeStackSize in non-Lowcode VMs.

RegisterAllocatingCogit:
Arrange to recompile when discovering loops with merges required to jump to the head.  Installing the register set at the end of the loop at its beginning means a much faster back branch on the common path and should be a win for simple loops.
Refactor resolveRegisterOrderConflictsBetweenCurrentSimStackAnd: into it and conflictingRegistersBetweenSimStackAnd: for use by ensureRegisterAssignmentsAreAtHeadOfLoop:.

Fix several slips where methodOrBlockNumArgs was used instead of methodOrBlockNumTemps.

Nuke unused method & variable.