Nuke the deadly openssl symlink
** Hack Hack Hack **
This is a horrible hack.
This is fragile because the mklink.pl file does not exist in 1.1.x version
But frankly, asking for a windows build to rely on symlink???? WTF???

I tried several unsuccessful workarounds
Configure no-symlinks
It does not work because then include/*.h are not built
I tried CYGWIN="winsymlinks" but it did not work (I don't even know if perl did take it into account)

With this, openssl almost build. Almost, because the generated makefile don't use the right cross-compilation windres...

    /bin/sh: gcc: command not found
    windres: preprocessing failed.
    i686-w64-mingw32-gcc: error: rc.o: No such file or directory