VMMaker.oscog-eem.2150 Stack/CoInterpreter: Refactor the assignments to the stack pointers from a page's head pointers into setStackPointersFromPage:.
Remove calls to addNewmethodToCache: from clients of lookupOrdinaryNoMNUEtcInClass: which does this anyway.

Cogit:
Use unsigned int as the type of 32-bit instructions (ARM & MIPS).  Reduce code on the class side in defining the CogAbstractInstruction typedefs.
Eliminate some "variable name shadows method" warnings.
Document that the limit on the Cogit's alloca more than ensures that the number of abstract instructions and bytecode fixups fits within 16 bits.
Make sure a simulation-only merthod is not generated.
FDix traceMerge: for in-image compilation.