CogVM source as per VMMaker.oscog-eem.2151
Stack/CoInterpreter:
Refactor the assignments to the stack pointers from a page's head pointers
into setStackPointersFromPage:.
Remove calls to addNewMethodToCache: from clients of
lookupOrdinaryNoMNUEtcInClass: which does this anyway.

Stack Interpreter Alien Callbacks:
Implement ownVM: and disownVM: for the non-threaded VM in StackInterpreter,
removing it from sqVirtualMachine.c. Arrange that sendInvokeCallbackContext:
& returnAs:ThroughCallback:Context: along with ownVM: and disownVM:
conspire to save and restore newMethod, argumentCount and
primitiveFunctionPointer around a callback. The VM depends on argumentCount
being correct to cut-back the correct number of arguments on primitive
return.  If a primitive that invokes a callback fails after invoking a
callback (a bad idea, but s**t happens during development) then newMethod
is required to activate the right faling method, and Spur expects
primitiveFunctionPointer to be valid, so asserts will fail misleadingly if
not.
Add three variables to help debug callbacks that are updated only in the
assert and debug VMs. debugCallbackPath has bits set in it that describe
the specific path taken through returnAs:ThroughCallback:Context:, while
debugCallbackReturns debugCallbackInvokes merely count invocations and
returns, and can be used for breakponts if reproducible cases arise (such
as the recent Pharo 6 VGTigerDemo runDemo issues with copyBits using
surface funciton callbacks.

64-bit Spur:
Fix variable declaration in primitiveSignalAtMilliseconds to prevent
arithmetic overflow in calculating nextWakeupUsecs on millisecond clock
rollover.  Addresses a problem reported for 32 bit images converted to 64
bit (e.g. Cuis port to Spur64) for which delay timing has not yet been
updated to use microsecond primitives.  Prior to this change, the following
would lock a 64 bit Spur image:
	s := Semaphore new.
	Delay primSignal: s atMilliseconds: Time primMillisecondClock - 10.
	s wait.

Sista
Fixed a bug where directed super send byte was leaking numExtB, leading to
the next instruction with ExtB being incorrectly interpreted

Spur64BitMemoryManager must export fetchClassTagOf:.

Cogit:
Use unsigned int as the type of 32-bit instructions (ARM & MIPS).  Reduce
code on the class side in defining the CogAbstractInstruction typedefs.
Document that the limit on the Cogit's alloca more than ensures that the
number of abstract instructions and bytecode fixups fits within 16 bits.

Rename blockPass, saveForBlockCompile et al to compilationPass,
saveForRecompile etc.  This so that RegisterAllocatingCogit can recompile
when discovering a loop that needs a merge so that the register set
computed at the end of a loop can be installed at the head.  Refactor
initializeFixupAt: into it & initializeFixup: to support RAC's recompilation.
Add some inlines to eliminate some thin intermediate functions.
Add support for adding #ifdef's to struct defs and use it to avoid
simNativeStackPtr and simNativeStackSize in non-Lowcode VMs.

Lowcode:
Refactoring the register allocation for the Lowcode instructions. The new
register allocators are not yet tested because an error on the previous
does not allow me to generate the sources for testing the refactoring.

Surface Plugin:
Allow defaulting the getSurfaceFormat, lockSurface, unlockSurface and
showSurface functions to 0 to save time if a subset of these are callbacks,
callbacks not being the fastest function invocations on the planet.

IA32ABIPlugin support files.
Rename abi support files to eliminate ia32 on non-ai32 platforms.

Slang:
Fix bug with type inference for #-.  The difference between two unsigned
values is now interpreted as signed.

Fix bug with inferring return types of methods that return unsigned typed
variables and positive integer constants.  Don't interpret the type of an
integer constant as #sqInt if the method also returns an unsigned.  So
defer considering the non-negative integer return values unless no other
return type information can be found.