InterpreterProxy: Add statNumGCs to answer the count of GCs so far in execution, used by primitives that may callback efficiently to detect the possible moving of objects.  Update vmProxyMinorVersion to 14 to reflect the new function.  Add PrimErrCallbackError to accompany PrimErrObjectMoved & PrimErrObjectNotPinned for informative error codes.
BitBlt plugin:
Check for GC after the query and lock callbacks, failing with PrimErrObjectMoved is so.  Reload source and dest forms in unlockSurface and showSurface if a GC has occurred, to ensure that the handles are accessed correctly.  Add numGCsOnInvocation bitBltIsReceiver inst vars to track GCs and the origin of the BitBlt oop so that reload works for the BalloonEnginePlugin too.  Add BEBitBltIndex class var for reloading in the context of a balloon primitive.  Nuke the unused JitBltHookSize class var.