CogVM source as per VMMaker.oscog-eem.2156
StackInterpreter:
Added extra condition in methodClassOf: to be able to ignore fullBlock last
literal without assertion failures.

InterpreterProxy:
Add statNumGCs to answer the count of GCs so far in execution, used by primitives
that may callback efficiently to detect the possible moving of objects.  Update
vmProxyMinorVersion to 14 to reflect the new function.  Add PrimErrCallbackError,
PrimErrObjectMoved & PrimErrObjectNotPinned for informative error codes.

BitBlt plugin:
Check for GC after the query and lock callbacks, failing with PrimErrObjectMoved
is so.  Reload source and dest forms in unlockSurface and showSurface if a GC has
occurred, to ensure that the handles are accessed correctly.  Add
numGCsOnInvocation bitBltIsReceiver inst vars to track GCs and the origin of the
BitBlt oop so that reload works for the BalloonEnginePlugin too.  Add
BEBitBltIndex class var for reloading in the context of a balloon primitive.

Sista:
Fix mapping back from inline cache tags to classes on 32-bit Spur where, because
SmallIntegers are 31 bits (tag = 1, not 01), Character (tag = 10) gets mapped to 0.

Cogit:
Fix the assert in allocateOpcodes:bytecodes:ifFail: by a) improving the sizeof:
estimates for CogBytecodeFixup and CogAbstractInstruction and b upping the factor
to multiply those sizes by.