CogVM source as per VMMaker.oscog-eem.2158
Fix stupid, stupid, stupid regression in VMMaker.oscog-eem.2151 where I added
savedPrimFunctionPointer to the VMCallbackContext structs in a position that
invalidates all the offsets in image code that refer to the elements of the rvs
(return value struct).  Adding savedPrimFunctionPointer is fine, but it has to
be /after/ the fields known to the image, which are all fields up to rvs.