CogVM source as per VMMaker.oscog-eem.2160
StackInterpreter:
Simplify the generated code for primitiveVMParameter by using beRootIfOld:
instead of storePointer: for the bulk store into the zero args case.  Make the
tenuringThreshold: setters use the same convention as other setters, answering
the primFailCode.

SpurCogit:
Fix regression in genNewHashTrampoline.

Fix the slip in creating the Spur ceNewHash trampoline, hence making the machine
code identityHash primitive non-failing.  Add a different trampoline for Sista
inline prim identityHash that saves all registers.

Sista:
Provide a movable allocation threshold in the mehtod zone so that nore space can
be made available on a counter trip to avoid reclaiming the method zone and
there-by destroying send and branch data as Scorch kicks in.  Allow the
threshold to be read and reset via vmParameterAt: 17.  Put this in the
SistaMethodZOne subclass of CogMethodZone.  Have the ceCounterTripped: routine
set the threshold to 1.0 from its default of 0.5.  Double the size of the
default code zone in Sista VMs.