CogVM source as per VMMaker.oscog-eem.2182
Fix pc mapping for backward jumps in Cogit VMs with extension bytecodes.  The
old code mapped a backward branch mcpc to the unextended jump bytecode, which
could turn a backward branch into a forward jump, with disastrous results.  The
map searcher must offset by the number of extensions to refer to the beginning
of the extended bytecode sequence for a backward jump.  The extended backward
jump bytecode must compute correctly the offset of the start of the extended
bytecode from the target of the jump.

Fix context printing so that the cogMethod is printed last in a stable context.

Sista Cogit:
Added and fixed a quite stable version of inlinedByteEqual.

Added experimental version of fillArray.

Temporary fix for inlined new.

Fix lowCode primitives in Sista VM.