CogVM source as per VMMaker.oscog-eem.2206
Fix the ARM regression introduced in VMMaker.oscog-eem.2160.  On ARM the link
register must be saved and restored around the call to the ceNewHash trampoline
in the identityHash primitive.

Initialize logging in the unix B3DAcceleratorPlugin, and add an error check t
to allow inferring from the log file that it has initialized correctly.

Add mesa-libGL-devel as to be installed in the linux HowToBuilds.