VMMaker.oscog-eem.2209.mcz Spur: Lift the limits on Spur image segments.  There is now no limit on the number of objects in a segment, but the classes in the segment must occupy no more than 2 ^ 21 bytes (maxIdentityHash / 2 * 8 - which allows for about 1.4 million 11 slot classes), and the segment can have no more than 2 ^ 21 out pointers (again maxIdentityHash / 2.  Tested for simple examples.  manages to export a project but not load it yet.
Fail if the segment store args that are mutated are either immutable or pinned.

Fix bad bug in shorten:toIndexableSize: (which could affect Scorch/Sista).  The old code for addressAfter: would be fooled by an object with overflow slows but fewer slots than numSlotsMask.  So fix addressAfter: so it will still add the overflow header if there is one.

Rewrite objectBytesForSlots: to follow the fixed addressAfter:.  The code should be cleaner.

Fix checkInterpreterIntegrity in the case of the primitiveDoPrimtiive code setting newMethod to an integer.

Fix some translation-timer warnings by tweaking storeLiteralVariable:withValue:, storePointerImmutabilityCheck:ofObject:withValue: & primitiveGetenv.

Simulator:
Discard the simulated implementations of makeDirEntryName:size:createDate:modDate:isDir:fileSize:posixPermissions:[isSymlink:] which are out of sync with the real code, simply defe4rring to the FilePlugin's actual methods.

Slang:
Eliminate halts in the conversion from Smalltalk parse tree to Slang parse tree to allow halts to be left in the simulator code without polluting the generated C.