VMMaker.oscog-nice.2214 Use a special form of sigsetjmp to avoid stack unwinding on Windows 64.
Note 1: stack unwinding is for C++ exception handling (for example for calling destructors).
We do not need it, and it requires the presence of specific information on stack which is not honoured by our JIT frames.

Note 2: this patch is specific to Mingw-w64. It might not work for another compiler (MSVC/Intel/...). Someone making the Vm compile with other tools might have to amend this #preambleCCode again.

In mingw, there is a possibility to pass a 2nd parameter to a _setjmp, and if this 2nd argument is NULL, there will be no context unwinding. This knowledge is extracted from these references:

http://mingw-w64-public.narkive.com/1mUoWEfG/setjmp-longjmp-crashes-second-setjmp-argument
https://patchwork.ozlabs.org/patch/437794/
http://www.agardner.me/golang/windows/cgo/64-bit/setjmp/longjmp/2016/02/29/go-windows-setjmp-x86.html
https://sourceforge.net/p/mingw-w64/bugs/465/

At current time of writing, _setjmp has 2 arguments #ifndef USE_NO_MINGW_SETJMP_TWO_ARGS 
and no header defines that by default:
grep -r USE_NO_MINGW_SETJMP_TWO_ARGS /usr/x86_64-w64-mingw32/sys-root/mingw/include