The code for restoring saved registers after a CallRT: did a benign iteration in excess.
The register save part does store the highBit of the mask in reg.
For example, for callerSavedRegsToBeSaved mask = 2r100,
we will have reg = 3 after the save loop.

The restore did start by inspecting bit 1<<reg, that is 2r1000, and that's one bit to far.
This is benign as long as no overflow occurs, but let's be safer.