CogVM source as per VMMaker.oscog-eem.2217
Cogit:
The code for restoring saved registers after a CallRT: did a benign iteration
in excess.

Spur:
Never inline objectsReachableFromRoots:.  Fix a speeling rorre.

SoundPlugin:
Add the Terf SoundPlugin extensions (to deal with multiple devices, their names,
volumes, etc).  Provide a nearly full implementation on Mac OS X Cocoa (inc 64
bits, but no AEC as yet).  Provide stub implementations for Unix & WIn32.  Any
one interested in win32/win64 I have the Terf C code in C++ files which needs
to be recast as C files.

Slang/FilePlugin:
Avoid remapping oops on creating directory entries in the FilePlugin on Spur by using remapOop:in:.  Change Slang to avoid creating nested #if-#else-#endif structures for nested remapOop:in:'s.  Nuke an unused configuration.

Cogit/CoInterpreter interaction:
Use a special form of sigsetjmp to avoid stack unwinding on Windows 64.

Note 1: stack unwinding is for C++ exception handling (for example for calling
destructors).  We do not need it, and it requires the presence of specific
information on stack which is not honoured by our JIT frames.

Note 2: this patch is specific to Mingw-w64. It might not work for another
compiler (MSVC/Intel/...). Someone making the Vm compile with other tools
might have to amend this #preambleCCode again.

In mingw, there is a possibility to pass a 2nd parameter to a _setjmp, and if
this 2nd argument is NULL, there will be no context unwinding. This knowledge
is extracted from these references:

http://mingw-w64-public.narkive.com/1mUoWEfG/setjmp-longjmp-crashes-second-setjmp-argument
https://patchwork.ozlabs.org/patch/437794/
http://www.agardner.me/golang/windows/cgo/64-bit/setjmp/longjmp/2016/02/29/go-windows-setjmp-x86.html
https://sourceforge.net/p/mingw-w64/bugs/465/

Build system:
Note that gcc beyond 4.8 does not work, at least on Debian linux.