VMMaker.oscog-nice.2220.mcz Fix display pixel reversal (for debug). Modern implementation of displayBits is (void *). We can't calculate an offset (dispalyBits + n) because we don't know the size of void! At least it fails on MSVC compiler. So it's necessary to convert it to char *, int *, or just sqInt.
While at it, as calculated, the range is a pixel range (assumed on the first display line) not a word range, but this range will be rounded to a whole word boundary.

The formulation now explicitely uses word indices, not word indices disguised in byte offsets.
It should avoid bad alignment cases where word(byte)StartIndex could eventually not be a proper multiple of 4.
It should also now work for any depth.
It might overflow for a pixel range highBit > 26 if sqInt were only 32 bits, that is over 64 million pixel width, but since our screens are not yet that wide, and will never be that wide in a 32 bits addressable space, we're safe.