VMMaker.oscog-eem.2222 StackInterpreter: Add time primitives to answer the UTC and local microsecond clocks as updated by the heartbeat (as the Cog VM was originally before David changed the primtiives to access the time now).  This access is much faster (of the order of 20 nsecs per invocartion vs 450 nsecs per invocatuon on a 2.3 GHz Mac Mini).  Terf demands very fast clock times but is not concerned with fine granularity.  The primitive numbers are 245 for UTC and 246 for local.
primitiveMakePoint should fail if its /argument/ is not a SmallInteger or Float.  Since the primitive is on Number we know the receiver is valid, and there is no point chekcing it.  Checking the argument allows us to e.g. make 3D point using a cascade of @'s.

Fix warnings for addIdleUsecs: on WIN32/64 (must not be marked exporrt to avoid declspec being added).

reverseDisplayFrom:to: slightly improved if it does the addition at start of loop, not on every iteration.

Cogit:
primitiveCollectCogCodeConstituents should verify its argument (if it has one) and fail if not given a boolean.  If it is given an argument it must remove it from the stack.

Correct comment in sendTrace:.