CogVM source as per VMMaker.oscog-eem.2222
StackInterpreter:
Add time primitives to answer the UTC and local microsecond clocks as updated
by the heartbeat (as the Cog VM was originally before David changed the
primitives to access the time now).  This access is much faster (of the order
of 20 nsecs per invocartion vs 450 nsecs per invocatuon on a 2.3 GHz Mac Mini).
Terf demands very fast clock times but is not concerned with fine granularity.
The primitive numbers are 245 for UTC and 246 for local.

primitiveMakePoint should fail if its /argument/ is not a SmallInteger or
Float.  Since the primitive is on Number we know the receiver is valid, and
there is no point checking it.

Fix warnings for addIdleUsecs: on WIN32/64 (must not be marked exporrt to
avoid declspec being added).

Fix display pixel reversal (for debug).  Modern implementation of displayBits
is (void *).  We can't calculate an offset (dispalyBits + n) because we don't
know the size of void! At least it fails on MSVC compiler.  So it's necessary
to convert it to char *, int *, or just sqInt.  While at it, as calculated,
the range is a pixel range (assumed on the first display line) not a word
range, but this range will be rounded to a whole word boundary.

Cogit:
Modified with Sophie K the collectCogConstituents primitive to take a boolean
parameter "withDetails". If false or no parameter, the primitive does the same
thing as before. If true, the primitive maps each CogMethod that is frameful
to an array starting with the cogMethod address followed by a mapping between
the mcpc and the bcpc instead of just the cogMethod address.

Fix the method for getting the stack pointer in Win64 ABI.  The generated
ceCaptureCStackPointers must account for the 32 bytes reserved for saving
the 4 register parameters which are allocated whatever the real number of
parameters.

Minor fix in register in branchIfInstanceOf: in SistaCogit for Characters.

Plugins:
Fix slips in the SoundPlugin code.

Win64 build.  Do not define WIN32.  Fix the win32 sqConfig.h file to work around
the clang preprocessor's failure to default undefined identifiers to 0L.