CogVM source as per VMMaker.oscog-eem.2227
Cogit:
Use genMarshallNArgs:arg:arg:arg:arg: when compiling an interpreter primitive
call to allow adding the shadow parameter space on Win64.

Remove a couple of unused and dubious methods in the CogAbstractInstruction
hierarchy.  Inline arg removal to match inlining of arg marshalling, since
removal is either no or one instruction.

Fix generation of JumpR on x64.  I forgot a rex byte.  This doesn't affect SysV
because we only ever jump through ClassReg (in an open PIC dispatch) which is
RCX (1) on SysV but R8 (8) on Win64.
With the fix to JumpR a Win64 image simulates correctly (with smashing of the
shadow parameter space).

FilePlugin:
Incorporate Monty's file plugin changes: add a new primitiveFileOpenNew to open an unexisting file atomically without race condition.