VMMaker.oscog-nice.2233 Integrate changes from monty to sqFileOpenNew.
There must be a way to check if failure is due to existence of file (implemented via PrimErrInappropriate).
If we would test existence afterward, the file could have been deleted in interim leading to another race condition.

This VMMaker change requires a coordinated platform source change (adding a 4th parameter to sqFileOpenNew).