CogVM source as per VMMaker.oscog-eem.2240
SistaCogit:
Fix (I think) genGetInstanceOfFixedClass:into:initializingIf: for duff's device
instance creation, and restore its use in genUnaryInlinePrimitive:.

Fixed a regression where primitiveCollectCogCodeConstituents would not extract
the details of CogMethods anymore if the first parameter is true.

Simulator:
Fix a bad bug with the simulation of the remembered set in 64-bits; the type
for rememberedSet was ending up as int *.  It must be int * on 32-bits but
long long * on 64-bits.