VMMaker.oscog-eem.2243 StackInterpreter: Make statTenures, statAverageLivePagesWhenMapping & statMaxPageCountWhenMapping (parameters 11, 68 & 69) writable to allow easier profiling.  Allow parameters expecting a float (statTenures, Sista CogCodeThreshold & statAverageLivePagesWhenMapping: 17, 55, 68) to take an int.
Slang:
Fix a bug in inferTypesForImplicitlyTypedVariablesIn:.  We cannot derive types from variables assigned to until all assignments are typed.  So exclude variables assigned from as-yet-untyped methods.  The old code would simply ignore the types from  as-yet-untyped methods, hence leaving the variable with a chosen-at-random, unmerged type.

This fixes a number of cases.  But there's still the weirdness that the return type of mapEndFor: in cogitX64SysV.c is correct (usqInt) but incorrect (sqInt) in cogitX64WIN64.c.  Luckily this is benign, but still should be fixed asap.