Spur: Round the space estimate when a new segment is used for planning compaction to avoid an assert fail.
Fix bad slip in computing objectsReachableFromRoots: (the wrong object was being marked when pushing unmarked objects onto the objStack), and rewrite to avoid growing the markStack unnecessarily.  Fix an assert fail in adding to the mark stack by handling totalFreeOldSpace correctly (and comment other callers of allocateLargestFreeChunk).

Refactor use of marked bit to identify heap objects copied into the segment into isCopiedIntoSegment: & markAsCopiedIntoSegment:.

Make sure unsigned comparisons are used for several address/oop comparisons in the segment code.

Add support for doing an image segment save and/or load in a clone (which was key in debugging the slip in objectsReachableFromRoots:).

Add some debugging routines: forwardersIn:, indexOf:in:.