CogVM source as per VMMaker.oscog-eem.2262
Modify the profiling primitive cogCodeConstituents: to be able to differentiate
the closedPICs from the openPICs in the profiling report

LargeIntegers plugin
Fix the crash for 2009 nthRoot: 100000 due to digitDivLarge:with:negative: not
checking if allocations fail.  The example produces 600k byte long integers and
so provokes plenty of allocation failures.

In addition mark some support methods as <inline: #always> to eliminate their
unnecessary uninlined versions.