CogVM source as per VMMaker.oscog-eem.2264
Spur:
The InterpreterProxy/sqVirtualMachine.h API (e.g. in the ThreadedFFIPlugin)
expects instantiateClass:indexableSize: to work with 0 indexable size for fixed
classes.  So remove the cppIf:... from the Spur implementation.

ObjectMemory:
Eliminate the header2 variable in a few routines.
It is always the same as classPointer.

General:
If ioScreenSize answers a zero screen extent (as is the case for headless
images) then on snapshot write the sawvedWindowSize rather thna the zero extent.
This should fix the experience people have with Pharo of saving headless images
that then come up with a zero sized window when started headful.

Image:
Fix a repository name in the image build script,