CogVM source as per VMMaker.oscog-eem.2265
Oops!  Writing the function does nothing unless it is called.  So once again,
if ioScreenSize answers a zero screen extent (as is the case for headless
images) then on snapshot write the sawvedWindowSize rather thna the zero extent.
This should fix the experience people have with Pharo of saving headless images
that then come up with a zero sized window when started headful.

Fix slip in iOS sqMacHostWindow.m.