CogVM source as per VMMaker.oscog-eem.2266
Spur:
Simplify cleverSwapHeaders:and:copyHashFlag: via xor.
Fix the type of currentAllocatedBytes given the Slang changes below.

Slang:
Fix a regression caused by the Slang fixes in VMMaker.oscog-eem.2243.
Since inferTypesForImplicitlyTypedVariablesIn: no longer sets the types of
locals as it goes along (which was incorrect) we can no longer default the types
of untyped variables to sqInt for the purposes of returnTypeForSend:in:ifNil:.
To allow returnTypeForSend:in:ifNil: to answer nil for arithmetic on untyped
expressions typeForArithmetic:in: uses TParseNode>>typeOrNilFrom:in: instead of
CCodeGenerator>>typeFor:in: to avoid the defaulting. returnTypeForSend:in:ifNil:
has been refactored to use the more direct TParseNode>>typeFrom:in: instead of
CCodeGenerator>>typeFor:in: for clarity.

The regression caused Slang to fail to infer the types of remembered1/2 &
hash1/2 in cleverSwapHeaders:and:copyHashFlag:.  The new code correctly
infers the types of remembered & hashBits as sqLong.

Plugins:
Eliminate a couple of warnings in the SoundPlugin by using 0 instead of NULL.

Fix paths in the win32 FT2Plugin makefile.