VMMaker.oscog-cb.2269.mcz Normally this commit has no impact on the production VM (some refactorings were made to share code between different part of the GC, but no new things). I tried to simulate the whole VM and generate a VM and it works, but only Travis can confirm us everything's ok. This is the first commit for the work done in the direction of a low pause GC. The work is split in two, an "engineering" part which consists in incremental marking derived from Dijkstra tri-color, and a "research" part which consists in selective compaction of the heap reusing forwarding objects.
This commit introduces SpurCompactor, an abstract class over the available compactors (Pig, Planning and the experimental ones). The full GC work is divided in Spur in two phases, the marking phase (incremental or not) and the compaction phase (Pig, planning or the experimental ones). 

The new compactor I want to build (SpurSelectiveCompactor) is a sweep algorithm which also compacts part of the heap. I started to do it in 2 steps. First I added SpurSweeper, a compactor which does not compact (if the programmer choose SpurSweeper as the compaction phase, the fullGC is a mark-sweep algorithm). Second I added SpurSelectiveCompactor, which reuses SpurSweeper to sweep the heap. We note that as these new compactors are not compacting the full heap, they should not be used for snapshots, hence snapshots use another compactor (currently SpurPlanningCompactor instead).