CogVM source as per VMMaker.oscog-eem.2277
Spur:
Fix the growth of the remembered table during GC when there is not enough
memory in old space. Allocates an extra memory segment in that case. This
is very uncommon.

Please check the thread: "[Vm-dev] VM crash with message
'could not grow remembered set' " to reproduce the problem.

Provide a logging facility that uses the Printf package.  See
VMClass>>log:with:* for API.  See Slang changes below for support.

Use the logging facility to log scavenges. Each scavenge produces output such as
scavenge 31 eden bytes: 0x2cfaf00/47165184 past bytes: 0x17c618/1558040
	remembered set: 73 redzone: 98304 size: 131072
	tenure below 0x0/0 by age
	survivor bytes: 0x17dd90/1564048 remembered objects: 73 tenured: 0
Use VM command line arg -/-logscavenger to enable.
This facility is provisional.  I (eliot) plan to reorganize so that
accurate timing can be included.

Cogit:

Added primitiveStringReplace in the JIT with Spur, implementing only the quick
paths for byte objects and array-like objects.

FFI:
Fixed bug with large struct in FFI with SysV x86-64.  A typo in the Smalltalk
source meant structs were processed incorrectly, confusing pass-in-register and
pass-on-stack cases.

Slang:
Add the notion of a varargs selector (currently anything ending in printf:, such as f:printf:).  Transform sends of such selectors with a brace of arguments into a varargs call (i.e. include all the arguments in the brace).

When inlining, if the method being inlined is of the form
	guard ifTrue: [...]
and the argument assignments are side-effect free,
and the guard does not refer to the arguments,
push the argument assignments past the guard into the block so that they too are guarded.